Make sure the lottery always draw from the set end block
Before, the choice of a random seed depended on which block the draw-numbers function was called.
This could lead to a situation where players on one side or the other could try to delay the number
drawing process, and wait for a favorable block before they call the function.

This commit addresses this issue by making sure the lottery will always pick the vrf-seed from the
lottery's end block minus one.